fib2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,14 +7,14 @@
 int main(void)
 {
 int count;
-unsigned long fib0 = 1, fib1 = 2, prevfeb;
+unsigned int fib0 = 1, fib1 = 2, prevfeb;
 
-printf("%lu, %lu", fib0, fib1);
+printf("%d, %d", fib0, fib1);
 for (count = 0; count < 96; count++)
 {
 prevfeb = fib1;
 fib1 = fib0 + fib1;
-printf(", %lu", fib1);
+printf(", %d", fib1);
 fib0 = prevfeb;
 }
 printf("\n");
